\name{FRaC-package}
\alias{FRaC-package}
\alias{FRaC}
\docType{package}
\title{Feature Regression and Classification Anomaly Detection Algorithm
}
\description{
FRaC is a general approach to the anomaly detection problem that is, the task of identifying instances that come from a different class or distribution than the majority (unsupervised anomaly detection) or a set of verified "normal" data (semi-supervised anomaly detection).The key to making this approach work is to precisely quantify the amount of evidence provided by each observation. To this end, FRaC developes a novel, information-theoretic anomaly measure that combines the contributions of all feature models.
}
\details{
\tabular{ll}{
Package: \tab FRaC\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2015-05-01\cr
License: \tab GPL 3.0\cr
}
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
K. Noto, C. E. Brodley, and D. Slonim.
FRaC: A Feature-Modeling Appraoch for Semi-Supervised and Unsupervised Anomaly Detection.
Data Mining and Knowledge Discovery, 25(1), pp.109—133, 2011.
}
\keyword{ package }


