% Generated by roxygen2 (4.0.2): do not edit by hand
\name{frac}
\alias{frac}
\title{FRaK: Feature Modeling Approach to Anomaly Detection}
\usage{
frac(x,  models, n.cv=2, keys = NULL grid=NULL, allowParallel = FALSE, tuneList = NULL)
}
\arguments{
\item{x}{An NxT matrix with T observations of N varialbles}

\item{models}{a string vector of models available in the caret package}

\item{keys}{A vector of integers representing places that the data set has time variables. This is only here for short term bug fixes}

\item{n.cv}{An integer specifying the number of cross-validations}

\item{allowParallel}{Logical definining whether or not the user would like to use a parallel backend if one is set up}

\item{tuneList}{A list of models for train functions}

\item{grid}{A list containing the different parameter values for each models iterations. Can be set to default NULL}
}
\value{
values or sup: the normalised suprisal score for each observation. Higher scores equate to a higher chance of an observation being an outlier
}
\description{
The FRaC Algorithm
}
\details{
Version 0.0.3 Alpha
}
\author{
Steve Bronder
}
\references{
K. Noto, C. E. Brodley, and D. Slonim.
FRaC: A Feature-Modeling Appraoch for Semi-Supervised and Unsupervised Anomaly Detection.
Data Mining and Knowledge Discovery, 25(1), pp.109—133, 2011.
}

